+++ packages/adapter-test/src/ambient.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="lucia-sveltekit" />
 declare namespace Lucia {
     type Auth = any;
-    type UserAttributesSchema = {
+    type UserAttributes = {
         username: string
     };
 }