+++ packages/lucia-sveltekit/src/ambient.d.ts
@@ -1,5 +1,5 @@
 declare namespace Lucia {
-    export type UserAttributesSchema = {};
+    export type UserAttributes = {};
     export class Auth extends (await import("./auth/index.js")).Auth {}
 }